fix: Use grey instead of blue in poll buttons & promise fix

+++ src/modules/polls.js
@@ -47,7 +47,7 @@ function generateButtonComponents(list, replaceLabel) {
     // eslint-disable-next-line no-restricted-syntax
     for (const key of Object.keys(list)) {
         components.push(new MessageButton()
-            .setStyle("PRIMARY")
+            .setStyle("SECONDARY")
             .setLabel(replaceLabel ?? list[key])
             .setEmoji(numberToEmoji(key).replace(/:/g, ""))
             .setCustomId(key))
@@ -97,7 +97,7 @@ async function createPoll(title, description, image, color, options, end, channe
  * End a poll
  * @param {Message} message
  * @param {Object} document
- * @returns {Promise<void>}
+ * @returns {Promise<Message<boolean>>}
  */
 async function endPoll(message, document) {
     message.embeds[0].fields[0].name = "Results"
@@ -107,7 +107,7 @@ async function endPoll(message, document) {
     message.embeds[0].fields = [message.embeds[0].fields[0]]
     message.embeds[0].setFooter("Poll ended.")
     message.embeds[0].timestamp = null
-    message.edit({
+    return message.edit({
         embeds: [message.embeds[0]], files: [], attachments: [], components: []
     })
 }
@@ -125,8 +125,15 @@ setInterval(async () => {
         }
         // Do we end the poll?
         if (document.end <= new Date().getTime()) {
-            await global.schemas.PollModel.findOneAndRemove({ id: document.id, message: document.message })
-            await endPoll(message, document)
+            let doRetry = true
+            try {
+                await endPoll(message, document) // This should throw and not update the database if this fails
+                doRetry = false
+                await global.schemas.PollModel.findOneAndRemove({ id: document.id, message: document.message })
+            } catch (err) {
+                // Try deletion once more
+                if (!doRetry) await global.schemas.PollModel.findOneAndRemove({ id: document.id, message: document.message })
+            }
             return
         }
         const buttonData = Object.fromEntries(Object.keys(document.votes).map((key) => [key, document.votes[key].length.toString()]))
